+++ sdk/resourcemanager/azure-resourcemanager-keyvault/src/main/java/com/azure/resourcemanager/keyvault/models/Vault.java
@@ -58,6 +58,11 @@ public interface Vault
     List<AccessPolicy> accessPolicies();
 
     /**
+     * @return whether role based access control (RBAC) for for authorization of data access is enabled.
+     */
+    boolean roleBasedAccessControlEnabled();
+
+    /**
      * @return whether Azure Virtual Machines are permitted to retrieve certificates stored as secrets from the key
      *     vault.
      */
@@ -148,6 +153,13 @@ public interface Vault
              * @return the first stage of the access policy definition
              */
             AccessPolicy.DefinitionStages.Blank<WithCreate> defineAccessPolicy();
+
+            /**
+             * Enables role based access control (RBAC) for authorization of data access.
+             *
+             * @return the next stage of key vault definition
+             */
+            WithCreate enableRoleBasedAccessControl();
         }
 
         /** A key vault definition allowing the networkAcl to be set. */
@@ -324,6 +336,20 @@ public interface Vault
              * @return the update stage of the access policy definition
              */
             AccessPolicy.Update updateAccessPolicy(String objectId);
+
+            /**
+             * Enables role based access control (RBAC) for authorization of data access.
+             *
+             * @return the update stage of the access policy definition
+             */
+            Update enableRoleBasedAccessControl();
+
+            /**
+             * Disables role based access control (RBAC) for authorization of data access.
+             *
+             * @return the update stage of the access policy definition
+             */
+            Update disableRoleBasedAccessControl();
         }
 
         /** A key vault update allowing the NetworkRuleSet to be set. */